pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-platfrom-deployer"
version = "1.2.0"
description = "CLI tool for automatic data platform deployment"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Khamidullin Ilsaf" },
]
license = { file = "LICENSE" }
dependencies = [
"absl-py==2.2.2",
"attrs==25.1.0",
"backports.tarfile==1.2.0",
"build==1.2.2.post1",
"certifi==2025.1.31",
"charset-normalizer==3.4.1",
"click==8.1.8",
"decorator==5.2.1",
"directory_tree==1.0.0",
"docutils==0.21.2",
"etils==1.12.2",
"fsspec==2025.3.2",
"id==1.5.0",
"idna==3.10",
"importlib_metadata==8.6.1",
"importlib_resources==6.5.2",
"jaraco.classes==3.4.0",
"jaraco.context==6.0.1",
"jaraco.functools==4.1.0",
"jsonpath-rw==1.4.0",
"jsonschema==4.23.0",
"jsonschema-specifications==2024.10.1",
"kagglehub==0.3.11",
"keyring==25.6.0",
"markdown-it-py==3.0.0",
"mdurl==0.1.2",
"more-itertools==10.6.0",
"directory-tree==1.0.0",
"networkx==3.4.2",
"nh3==0.2.21",
"packaging==24.2",
"ply==3.11",
"psycopg2-binary==2.9.10",
"Pygments==2.19.1",
"pyparsing==3.2.3",
"pyproject_hooks==1.2.0",
"python-dateutil==2.9.0.post0",
"pytz==2025.2",
"PyYAML==6.0.2",
"rdflib==7.1.4",
"readme_renderer==44.0",
"referencing==0.36.2",
"requests==2.32.3",
"requests-toolbelt==1.0.0",
"rfc3986==2.0.0",
"rich==13.9.4",
"rpds-py==0.23.1",
"scipy==1.15.2",
"six==1.17.0",
"SQLAlchemy==2.0.40",
"tqdm==4.67.1",
"twine==6.1.0",
"types-pytz==2025.2.0.20250326",
"typing_extensions==4.12.2",
"tzdata==2025.2",
"urllib3==2.3.0",
"zipp==3.21.0"
]

[tool.setuptools.package-data]
"dpd" = ["*.json","services/postgresql/postgresql.conf"]

[project.urls]
Repository = "https://github.com/Ilsaffff/data-platform-deployer"


[project.scripts]
dpd = "dpd.main:main"